v4
Version 4.0

+++ app/src/main/java/com/bluesoftit/cashcoin/Models/WithdrawRequest.java
@@ -44,7 +44,7 @@ public class WithdrawRequest {
         this.requestedBy = requestedBy;
     }
 
-    public long getWithdrawAmmount() {
+    public double getWithdrawAmmount() {
         return withdrawAmmount;
     }
